Split api and system controller
